Reorganize component and add jsdocs
